Unknown version

+++ ContraAI/Code/util.cpp
@@ -20,7 +20,7 @@ bool Stack::isEmpty(){
 	return false;
 }
 
-int manhattanDist(DOUBLE2 p1, DOUBLE2 p2){
+double manhattanDist(DOUBLE2 p1, DOUBLE2 p2){
 	return abs(p1.x - p2.x) + abs(p1.y - p2.y);
 }
